+++ ft_atoi_base.c
@@ -6,7 +6,7 @@
 /*   By: esterna <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/12 18:37:52 by esterna           #+#    #+#             */
-/*   Updated: 2017/03/12 20:22:03 by esterna          ###   ########.fr       */
+/*   Updated: 2017/10/22 18:33:22 by esterna          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,11 +30,11 @@ static int					ind(int n, const char *base)
 	return (i);
 }
 
-int							ft_atoi_base(const char *n, int b)
+size_t						ft_atoi_base(const char *n, int b)
 {
-	int				result;
-	int				neg;
-	int				len;
+	long long		result;
+	long long		neg;
+	long long		len;
 	const char		*base;
 
 	result = 0;
